Add archive thread back to context
